+++ core/src/main/java/org/infinispan/container/DataContainer.java
@@ -6,6 +6,8 @@ import java.util.Spliterators;
 
 import org.infinispan.container.entries.InternalCacheEntry;
 import org.infinispan.container.impl.InternalEntryFactory;
+import org.infinispan.eviction.impl.ActivationManager;
+import org.infinispan.eviction.impl.PassivationManager;
 import org.infinispan.factories.scopes.Scope;
 import org.infinispan.factories.scopes.Scopes;
 import org.infinispan.metadata.Metadata;
@@ -56,7 +58,7 @@ public interface DataContainer<K, V> extends Iterable<InternalCacheEntry<K, V>> 
     * Puts an entry in the cache along with metadata adding information such lifespan of entry, max idle time, version
     * information...etc.
     * <p/>
-    * The {@code key} must be activate by invoking {@link org.infinispan.eviction.ActivationManager#activateAsync(Object, int)}
+    * The {@code key} must be activate by invoking {@link ActivationManager#activateAsync(Object, int)}
     * boolean)}.
     *
     * @param k key under which to store entry
@@ -76,7 +78,7 @@ public interface DataContainer<K, V> extends Iterable<InternalCacheEntry<K, V>> 
    /**
     * Removes an entry from the cache
     * <p/>
-    * The {@code key} must be activate by invoking {@link org.infinispan.eviction.ActivationManager#activateAsync(Object, int)}
+    * The {@code key} must be activate by invoking {@link ActivationManager#activateAsync(Object, int)}
     *
     * @param k key to remove
     * @return entry removed, or null if it didn't exist or had expired
@@ -112,7 +114,7 @@ public interface DataContainer<K, V> extends Iterable<InternalCacheEntry<K, V>> 
    /**
     * Atomically, it removes the key from {@code DataContainer} and passivates it to persistence.
     * <p/>
-    * The passivation must be done by invoking the method {@link org.infinispan.eviction.PassivationManager#passivateAsync(InternalCacheEntry)}.
+    * The passivation must be done by invoking the method {@link PassivationManager#passivateAsync(InternalCacheEntry)}.
     *
     * @param key The key to evict.
     */
@@ -124,7 +126,7 @@ public interface DataContainer<K, V> extends Iterable<InternalCacheEntry<K, V>> 
     * See {@link org.infinispan.container.DataContainer.ComputeAction#compute(Object,
     * org.infinispan.container.entries.InternalCacheEntry, InternalEntryFactory)}.
     * <p/>
-    * The {@code key} must be activated by invoking {@link org.infinispan.eviction.ActivationManager#activateAsync(Object, int)}.
+    * The {@code key} must be activated by invoking {@link ActivationManager#activateAsync(Object, int)}.
     * <p>
     * Note the entry provided to {@link org.infinispan.container.DataContainer.ComputeAction} may be expired as these
     * entries are not filtered as many other methods do.
