+++ core/src/main/java/org/infinispan/eviction/impl/ActivationManagerStub.java
@@ -2,7 +2,6 @@ package org.infinispan.eviction.impl;
 
 import java.util.concurrent.CompletionStage;
 
-import org.infinispan.eviction.ActivationManager;
 import org.infinispan.factories.annotations.SurvivesRestarts;
 import org.infinispan.factories.scopes.Scope;
 import org.infinispan.factories.scopes.Scopes;
@@ -14,14 +13,6 @@ import org.infinispan.util.concurrent.CompletableFutures;
 @Scope(Scopes.NAMED_CACHE)
 @SurvivesRestarts
 public class ActivationManagerStub implements ActivationManager {
-   @Override
-   public void onUpdate(Object key, boolean newEntry) {
-   }
-
-   @Override
-   public void onRemove(Object key, boolean newEntry) {
-   }
-
    @Override
    public CompletionStage<Void> activateAsync(Object key, int segment) {
       return CompletableFutures.completedNull();
