+++ core/src/main/java/org/infinispan/eviction/impl/PassivationManagerImpl.java
@@ -8,7 +8,6 @@ import java.util.concurrent.CompletionStage;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
 
-import org.infinispan.commons.CacheException;
 import org.infinispan.commons.time.TimeService;
 import org.infinispan.commons.util.IteratorMapper;
 import org.infinispan.commons.util.Util;
@@ -19,7 +18,6 @@ import org.infinispan.container.impl.InternalDataContainer;
 import org.infinispan.context.impl.ImmutableContext;
 import org.infinispan.distribution.DistributionManager;
 import org.infinispan.distribution.ch.KeyPartitioner;
-import org.infinispan.eviction.AbstractPassivationManager;
 import org.infinispan.factories.annotations.Inject;
 import org.infinispan.factories.annotations.Start;
 import org.infinispan.notifications.cachelistener.CacheNotifier;
@@ -39,11 +37,11 @@ public class PassivationManagerImpl extends AbstractPassivationManager {
    private static final boolean trace = log.isTraceEnabled();
 
    @Inject PersistenceManager persistenceManager;
-   @Inject CacheNotifier notifier;
+   @Inject CacheNotifier<Object, Object> notifier;
    @Inject Configuration cfg;
    @Inject InternalDataContainer<Object, Object> container;
    @Inject TimeService timeService;
-   @Inject MarshallableEntryFactory marshalledEntryFactory;
+   @Inject MarshallableEntryFactory<?, ?> marshalledEntryFactory;
    @Inject DistributionManager distributionManager;
    @Inject KeyPartitioner keyPartitioner;
 
@@ -74,9 +72,9 @@ public class PassivationManagerImpl extends AbstractPassivationManager {
       return distributionManager != null && !distributionManager.getCacheTopology().isWriteOwner(key);
    }
 
-   private CompletionStage<Boolean> doPassivate(Object key, CacheEntry entry) {
+   private CompletionStage<Boolean> doPassivate(Object key, CacheEntry<?, ?> entry) {
       if (trace) log.tracef("Passivating entry %s", toStr(key));
-         MarshallableEntry marshalledEntry = marshalledEntryFactory.create(key, entry.getValue(), entry.getMetadata(),
+         MarshallableEntry<?, ?> marshalledEntry = marshalledEntryFactory.create(key, entry.getValue(), entry.getMetadata(),
                entry.getCreated(), entry.getLastUsed());
          CompletionStage<Void> stage = passivationPersistenceManager.passivate(marshalledEntry, keyPartitioner.getSegment(key));
          return stage.handle((v, t) -> {
@@ -92,28 +90,7 @@ public class PassivationManagerImpl extends AbstractPassivationManager {
    }
 
    @Override
-   public void passivate(InternalCacheEntry entry) {
-      Object key;
-      if (enabled && entry != null && !isL1Key(key = entry.getKey())) {
-         // notify listeners that this entry is about to be passivated
-         CompletionStages.join(notifier.notifyCacheEntryPassivated(key, entry.getValue(), true,
-               ImmutableContext.INSTANCE, null));
-         if (trace) log.tracef("Passivating entry %s", toStr(key));
-         try {
-            MarshallableEntry marshalledEntry = marshalledEntryFactory.create(key, entry.getValue(), entry.getMetadata(),
-                  entry.getExpiryTime(), entry.getLastUsed());
-            persistenceManager.writeToAllNonTxStoresSync(marshalledEntry, keyPartitioner.getSegment(key), BOTH);
-            if (statsEnabled) passivations.getAndIncrement();
-         } catch (CacheException e) {
-            CONTAINER.unableToPassivateEntry(key, e);
-         }
-         CompletionStages.join(notifier.notifyCacheEntryPassivated(key, null, false,
-               ImmutableContext.INSTANCE, null));
-      }
-   }
-
-   @Override
-   public CompletionStage<Void> passivateAsync(InternalCacheEntry entry) {
+   public CompletionStage<Void> passivateAsync(InternalCacheEntry<?, ?> entry) {
       Object key;
       if (enabled && entry != null && !isL1Key(key = entry.getKey())) {
          if (notifier.hasListener(CacheEntryPassivated.class)) {
@@ -129,16 +106,7 @@ public class PassivationManagerImpl extends AbstractPassivationManager {
 
    @Override
    public void passivateAll() throws PersistenceException {
-      if (enabled && !skipOnStop) {
-         long start = timeService.time();
-         CONTAINER.passivatingAllEntries();
-
-         int count = container.sizeIncludingExpired();
-         Iterable<MarshallableEntry> iterable = () -> new IteratorMapper<>(container.iterator(), e ->
-            marshalledEntryFactory.create(e.getKey(), e.getValue(), e.getMetadata(), e.getExpiryTime(), e.getLastUsed()));
-         CompletionStages.join(persistenceManager.writeBatchToAllNonTxStores(iterable, BOTH, 0));
-         CONTAINER.passivatedEntries(count, Util.prettyPrintTime(timeService.timeDuration(start, TimeUnit.MILLISECONDS)));
-      }
+      CompletionStages.join(passivateAllAsync());
    }
 
    @Override
