+++ core/src/main/java/org/infinispan/factories/EmptyConstructorNamedCacheFactory.java
@@ -22,9 +22,9 @@ import org.infinispan.distribution.L1Manager;
 import org.infinispan.distribution.RemoteValueRetrievedListener;
 import org.infinispan.distribution.TriangleOrderManager;
 import org.infinispan.distribution.impl.L1ManagerImpl;
-import org.infinispan.eviction.ActivationManager;
+import org.infinispan.eviction.impl.ActivationManager;
 import org.infinispan.eviction.EvictionManager;
-import org.infinispan.eviction.PassivationManager;
+import org.infinispan.eviction.impl.PassivationManager;
 import org.infinispan.eviction.impl.ActivationManagerImpl;
 import org.infinispan.eviction.impl.EvictionManagerImpl;
 import org.infinispan.eviction.impl.PassivationManagerImpl;
