+++ server/runtime/src/test/java/org/infinispan/server/unit/configuration/ServerEventLoggerTest.java
@@ -18,7 +18,7 @@ import org.infinispan.configuration.cache.CacheMode;
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.configuration.global.GlobalConfiguration;
 import org.infinispan.configuration.global.GlobalConfigurationBuilder;
-import org.infinispan.eviction.PassivationManager;
+import org.infinispan.eviction.impl.PassivationManager;
 import org.infinispan.manager.EmbeddedCacheManager;
 import org.infinispan.persistence.file.SingleFileStore;
 import org.infinispan.server.logging.events.ServerEventImpl;
@@ -30,6 +30,7 @@ import org.infinispan.test.MultiCacheManagerCallable;
 import org.infinispan.test.TestingUtil;
 import org.infinispan.test.fwk.TestCacheManagerFactory;
 import org.infinispan.test.fwk.TestResourceTracker;
+import org.infinispan.util.concurrent.CompletionStages;
 import org.infinispan.util.logging.events.EventLog;
 import org.infinispan.util.logging.events.EventLogCategory;
 import org.infinispan.util.logging.events.EventLogLevel;
@@ -148,7 +149,8 @@ public class ServerEventLoggerTest {
          EventLogger eventLogger = EventLogManager.getEventLogger(cm);
          eventLogger.info(EventLogCategory.CLUSTER, "message #1");
          Cache<UUID, ServerEventImpl> cache = cm.getCache(ServerEventLogger.EVENT_LOG_CACHE);
-         TestingUtil.extractComponent(cache, PassivationManager.class).passivateAll();
+         PassivationManager passivationManager = TestingUtil.extractComponent(cache, PassivationManager.class);
+         CompletionStages.join(passivationManager.passivateAllAsync());
          SingleFileStore<UUID, ServerEventImpl> sfs = TestingUtil.getFirstWriter(cache);
          assertEquals(1, sfs.size());
       });
