+++ wildfly/infinispan/src/main/java/org/jboss/as/clustering/infinispan/subsystem/CacheCommands.java
@@ -30,7 +30,7 @@ import java.util.ListIterator;
 import java.util.Map;
 
 import org.infinispan.Cache;
-import org.infinispan.eviction.PassivationManager;
+import org.infinispan.eviction.impl.PassivationManager;
 import org.infinispan.factories.ComponentRegistry;
 import org.infinispan.interceptors.impl.CacheLoaderInterceptor;
 import org.infinispan.interceptors.impl.CacheMgmtInterceptor;
